A very simple SDK (subset) for the LINE Messaging API for PHP.
line-bot-sdk-php is a full-stack implementation of the LINE Messaging
API SDK, which uses an OOP interface and functions. It provides an API
client, a message builder, an HTTP client, an event parser and other
useful components.

On the other hand, line-bot-sdk-tiny provides a simple interface and
functions. It contains a part of the API functions (not full function).

This SDK contains only one file, so it is easy to add LINEBotTiny.php to
your environment and require that from your script. And of course, you
can also copy and paste this SDK.